refactor: fold Core into Infrastructure hosting, runtime and UI services
