+++ OLED-Sleeper.Tests/UI/ViewModels/MainViewModelTests.cs
@@ -1,5 +1,4 @@
 using Moq;
-using OLED_Sleeper.Core.Interfaces;
 using OLED_Sleeper.Features.MonitorInformation.Models;
 using OLED_Sleeper.Features.UserSettings.Models;
 using OLED_Sleeper.Features.UserSettings.Services.Interfaces;
