+++ OLED-Sleeper/Infrastructure/Hosting/ApplicationBootstrapper.cs
@@ -1,14 +1,15 @@
 using Microsoft.Extensions.DependencyInjection;
-using OLED_Sleeper.Core;
-using OLED_Sleeper.Core.Interfaces;
 using OLED_Sleeper.Infrastructure.Helpers;
+using OLED_Sleeper.Infrastructure.Hosting.Interfaces;
+using OLED_Sleeper.Infrastructure.Runtime;
+using OLED_Sleeper.Infrastructure.Runtime.Interfaces;
 using OLED_Sleeper.UI.Services.Interfaces;
 using Serilog;
 using System;
 using System.Linq;
 using System.Windows;
 
-namespace OLED_Sleeper.Infrastructure
+namespace OLED_Sleeper.Infrastructure.Hosting
 {
     /// <summary>
     /// Handles application startup, dependency injection, single-instance enforcement, orchestrator startup, and shutdown logic.
