+++ OLED-Sleeper/Infrastructure/Hosting/Interfaces/IApplicationInstanceManager.cs
@@ -1,4 +1,4 @@
-namespace OLED_Sleeper.Core.Interfaces
+namespace OLED_Sleeper.Infrastructure.Hosting.Interfaces
 {
     /// <summary>
     /// Defines the contract for managing single-instance enforcement and inter-process signaling.
