+++ OLED-Sleeper/UI/Services/MainWindowAccessor.cs
@@ -1,8 +1,8 @@
-using OLED_Sleeper.Core.Interfaces;
+using OLED_Sleeper.UI.Services.Interfaces;
 using System.Diagnostics.CodeAnalysis;
 using System.Windows;
 
-namespace OLED_Sleeper.Infrastructure
+namespace OLED_Sleeper.UI.Services
 {
     /// <summary>
     /// Reaches the main window through <see cref="Application.MainWindow"/>.
