+++ OLED-Sleeper/UI/ViewModels/MainViewModel.cs
@@ -1,6 +1,6 @@
-﻿using OLED_Sleeper.Core.Interfaces;
-using OLED_Sleeper.Features.MonitorDimming.Commands;
+﻿using OLED_Sleeper.Features.MonitorDimming.Commands;
 using OLED_Sleeper.Features.UserSettings.Services.Interfaces;
+using OLED_Sleeper.Infrastructure.Runtime.Interfaces;
 using OLED_Sleeper.Messaging.Interfaces;
 using OLED_Sleeper.UI.Commands;
 using OLED_Sleeper.UI.Helpers;
